requirements.txt
streamlit
openai
python-docx
pdfminer.six
python-dotenv
google-auth
google-api-python-client
tiktoken

